meta-ibm: Add Genesis3 support
This adds support for the IBM Genesis3 board.

It's an evaluation platform based on Aspeed AST2500 SoC powering
a 4 socket Intel Cascade Lake server. The main difference to regular
platforms is the non-existence of a CPLD/FPGA to do the power sequencing.
Everything on this board is directly connected to and controlled by the
BMC.

The following features are implemented:
- sensors
- fan control
- enabling/disabling voltage regulators
- powering the x86 system from G3 to S0 using pwrseqd (a software
  implementation of what is normally done with a CPLD on server
  mainboards).

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>
Change-Id: Ia040dca83fbe17c52d9aa87cbf45d90452c031ff