[NOT FOR MERGE]meta-ibm: Add custom kernel for SBP1
This commit must not be merged but helps developers testing the code for the
SBP1 platform until everything has been upstreamed.

Includes drivers and DTS to get this running.
Some drivers are in the upstream process, some are upstreamed by 3rd parties.

List of required drivers:
- Aspeed eSPI
- Aspeed JTAG
- bcm6123
- cy8c95x0
- ir38263
- infineon xdpe
- regulator core
- Aspeed PECI
- max735x
- max6639
- max597x
- PMBUS core

Already merged into Linux 6.1:
- bcm6123
- Aspeed PECI
- cy8c95x0
- ir38263
- max597x 1/2
- infineon xdpe

Change-Id: Icc5c23e6141ac4f2448c54736d6d8fd47ca10b43
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>