[NOT FOR MERGE]meta-ibm: Add custom power sequencing
The Genesis3 has no CPLD and thus all x86 power managment must be done
in software. Hook up pwrseqd to take control of 112 GPIOs, 80 voltage regulators
and let it do the power sequencing as defined in configuration files.

The Linux kernel must be modified to support:
- sysfs notify events on regulator changes
- allow to control regulators from userspace

This was built for openbmc, duplicates the behavior and features of
x86-power-control and violates the guidelines:

https://github.com/openbmc/docs/blob/master/meta-layer-guidelines.md#meta-layers-should-not-point-to-openbmc-specific-repositories-outside-of

TODO: Palce under OpenBMC umbrella.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Change-Id: Ic8ef227b5a3f657c710838296a608ea1835053a9