[NOT FOR MERGE]meta-ibm: Add custom power sequencing
The SBP1 has no CPLD and thus all x86 power managment must be done
in software. Hook up pwrseqd to take control of 226 GPIOs, 80 voltage regulators
and let it do the power sequencing as defined in configuration files.

The Linux kernel must be modified to support:
- sysfs notify events on regulator changes
- allow to control regulators from userspace

This was built for openbmc, duplicates the behavior and features of
x86-power-control and violates the guidelines:

https://github.com/openbmc/docs/blob/master/meta-layer-guidelines.md#meta-layers-should-not-point-to-openbmc-specific-repositories-outside-of

TODO: Palce under OpenBMC umbrella.

Change-Id: I8262af723dcd55c5fb6c1e962a54c4de5e9c3cc3
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>