ibm/meta-sbp1: Add bios version package
This package update bios version flashed in x86 PCH SPI flash.
It depends on the chassis to be powered and the Intel SPS firmware to
be loaded.

The bios version string is update over dbus & a copy of it is also
cached for later use to avoid reading SPI flash sibsequently.

Change-Id: I20a33e505bf6df82537cae67622f868e2a9b5a83
Signed-off-by: Naresh Solanki <Naresh.Solanki@9elements.com>